feat: Implement event management API with GET, DELETE, and PUT methods
style: Add styles for logout button and event adder components

feat: Create reusable Select component with Shadcn UI

feat: Add event card and upcoming events components with styles

feat: Implement hooks for adding, deleting, and updating events

feat: Create room management hooks and utility functions

+++ src/app/api/companies/[companyId]/route.tsx
@@ -1,3 +1,4 @@
+import { getUserProfileFromRequest } from "@/lib/auth/get_user_profile";
 import { createClient } from "@/lib/supabase/server";
 import { NextRequest } from "next/server";
 
@@ -5,6 +6,14 @@ export async function GET(
   request: NextRequest,
   { params }: { params: Promise<{ companyId: string }> }
 ) {
+  const userProfile = await getUserProfileFromRequest();
+
+  if (!userProfile || !userProfile.role || userProfile.role !== "admin") {
+    return new Response(JSON.stringify({ error: "Unauthorized" }), {
+      status: 401,
+    });
+  }
+
   const id = (await params).companyId;
   const supabase = await createClient();
   const { data, error } = await supabase
@@ -16,19 +25,19 @@ export async function GET(
   if (error) {
     console.error("Error fetching company:", error);
     return new Response(JSON.stringify({ error: "Internal server error." }), {
-      status: 500
+      status: 500,
     });
   }
 
   if (!data) {
     return new Response(JSON.stringify({ error: "Company not found." }), {
-      status: 404
+      status: 404,
     });
   }
 
   return new Response(JSON.stringify(data), {
     status: 200,
-    headers: { "Content-Type": "application/json" }
+    headers: { "Content-Type": "application/json" },
   });
 }
 
@@ -36,6 +45,14 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: Promise<{ companyId: string }> }
 ) {
+  const userProfile = await getUserProfileFromRequest();
+
+  if (!userProfile || !userProfile.role || userProfile.role !== "admin") {
+    return new Response(JSON.stringify({ error: "Unauthorized" }), {
+      status: 401,
+    });
+  }
+
   const id = (await params).companyId;
   const supabase = await createClient();
   const { error } = await supabase.from("companies").delete().eq("id", id);
@@ -43,7 +60,7 @@ export async function DELETE(
   if (error) {
     console.error("Error deleting company:", error);
     return new Response(JSON.stringify({ error: "Internal server error." }), {
-      status: 500
+      status: 500,
     });
   }
 
@@ -51,7 +68,7 @@ export async function DELETE(
     JSON.stringify({ message: "Company deleted successfully." }),
     {
       status: 200,
-      headers: { "Content-Type": "application/json" }
+      headers: { "Content-Type": "application/json" },
     }
   );
 }
@@ -60,6 +77,14 @@ export async function PUT(
   request: NextRequest,
   { params }: { params: Promise<{ companyId: string }> }
 ) {
+  const userProfile = await getUserProfileFromRequest();
+
+  if (!userProfile || !userProfile.role || userProfile.role !== "admin") {
+    return new Response(JSON.stringify({ error: "Unauthorized" }), {
+      status: 401,
+    });
+  }
+
   const id = (await params).companyId;
 
   const supabase = await createClient();
@@ -75,18 +100,18 @@ export async function PUT(
   if (error) {
     console.error("Error updating company:", error);
     return new Response(JSON.stringify({ error: "Internal server error." }), {
-      status: 500
+      status: 500,
     });
   }
 
   if (!data) {
     return new Response(JSON.stringify({ error: "Company not found." }), {
-      status: 404
+      status: 404,
     });
   }
 
   return new Response(JSON.stringify(data), {
     status: 200,
-    headers: { "Content-Type": "application/json" }
+    headers: { "Content-Type": "application/json" },
   });
 }
